requirements.txt
streamlit
requests
pydantic
instructor
openai
python-dotenv
streamlit_audiorecorder 